pyproject.toml
[tool.poetry]
name = "menugen"
version = "0.1.0"
description = "Generate FreeDesktop XDG menu entries for programs installed in the user home directory"
authors = ["'Fernando López' <'fernando.e.lopez@gmail.com'>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
python-magic = "^0.4.24"
xdg = "^5.1.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
pytest = "^6.2.4"
pre-commit = "^2.13.0"
pytest-mock = "^3.6.1"
pytest-watch = "^4.2.0"
parameterized = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
